requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
jinja2
requests
openai
python-dotenv
moviepy
imageio
imageio-ffmpeg
Pillow
psycopg2-binary
gunicorn
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-docx
pypdf
reportlab
mercadopago
gTTS
numpy
apscheduler
passlib[bcrypt]
python-jose[cryptography]
